Moved MinMax function to util and added testcases.
